+++ src/components/index.ts
@@ -1,3 +1,5 @@
 export * from "./Experience"
+export * from "./ExternalLinkButton"
+export * from "./FindMeLink"
 export * from "./icons"
 export * from "./SkillButton"
